add accessible view provider
