+++ src/vs/workbench/contrib/chat/browser/chat.ts
@@ -4,11 +4,14 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { Event } from 'vs/base/common/event';
+import { IDisposable, toDisposable } from 'vs/base/common/lifecycle';
 import { URI } from 'vs/base/common/uri';
 import { ICodeEditor } from 'vs/editor/browser/editorBrowser';
 import { Selection } from 'vs/editor/common/core/selection';
+import { registerSingleton, InstantiationType } from 'vs/platform/instantiation/common/extensions';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
 import { IChatWidgetContrib } from 'vs/workbench/contrib/chat/browser/chatWidget';
+import { ICodeBlockActionContext } from 'vs/workbench/contrib/chat/browser/codeBlockPart';
 import { IChatAgentCommand, IChatAgentData } from 'vs/workbench/contrib/chat/common/chatAgents';
 import { IParsedChatRequest } from 'vs/workbench/contrib/chat/common/chatParserTypes';
 import { IChatRequestViewModel, IChatResponseViewModel, IChatViewModel, IChatWelcomeMessageViewModel } from 'vs/workbench/contrib/chat/common/chatViewModel';
@@ -16,6 +19,7 @@ import { IChatRequestViewModel, IChatResponseViewModel, IChatViewModel, IChatWel
 export const IChatWidgetService = createDecorator<IChatWidgetService>('chatWidgetService');
 export const IQuickChatService = createDecorator<IQuickChatService>('quickChatService');
 export const IChatAccessibilityService = createDecorator<IChatAccessibilityService>('chatAccessibilityService');
+export const ICodeBlockContextProviderRegistry = createDecorator<ICodeBlockContextProviderRegistry>('codeBlockContextProviderRegistry');
 
 export interface IChatWidgetService {
 
@@ -134,3 +138,23 @@ export interface IChatWidget {
 export interface IChatViewPane {
 	clear(): void;
 }
+
+
+export interface ICodeBlockActionContextProvider {
+	getCodeBlockContext(editor?: ICodeEditor): ICodeBlockActionContext | undefined;
+}
+export interface ICodeBlockContextProviderRegistry {
+	serviceBrand: undefined;
+}
+export class CodeBlockContextProviderRegistry implements ICodeBlockContextProviderRegistry {
+	serviceBrand: undefined;
+	static readonly _providers = new Map<string, ICodeBlockActionContextProvider>();
+	static get providers(): ICodeBlockActionContextProvider[] {
+		return [...this._providers.values()];
+	}
+	static registerProvider(provider: ICodeBlockActionContextProvider, id: string): IDisposable {
+		this._providers.set(id, provider);
+		return toDisposable(() => this._providers.delete(id));
+	}
+}
+registerSingleton(ICodeBlockContextProviderRegistry, CodeBlockContextProviderRegistry, InstantiationType.Delayed);
